requirements.txt
anyio==3.6.1
black==22.8.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
docutils==0.19
fastapi==0.82.0
flit==3.7.1
flit_core==3.7.1
graphql-core==3.2.1
greenlet==1.1.3
h11==0.13.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.5.0
python-dateutil==2.8.2
python-multipart==0.0.5
python-slugify==6.1.2
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
starlette==0.19.1
strawberry-graphql==0.128.0
text-unidecode==1.3
tomli==2.0.1
tomli_w==1.0.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
